Project List now animates when blurb is shown
Changes to be committed:
	modified:   src/App/App.jsx
	renamed:    src/App/components/Title/Title.css -> src/App/components/Home/Home.css
	new file:   src/App/components/Home/Home.jsx
	modified:   src/App/components/Project-List/Project-list.css
	modified:   src/App/components/Project-List/Project-list.jsx
	deleted:    src/App/components/Title/Title.jsx
