+++ src/App/components/Project-List/Project-list.css
@@ -11,8 +11,9 @@
     
     width: fit-content;
     padding: 1rem;
-
-    outline: auto;
+    margin: 1rem;
+    border: 1px solid black;
+    
 }
 
 #projects-list {
@@ -25,7 +26,5 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    
-    /* outline: 1px solid red; */
 }
