+++ src/App/components/Project-List/Project-list.jsx
@@ -1,23 +1,24 @@
 import './Project-list.css'
 import { Link } from 'react-router-dom'
+import { motion } from 'framer-motion'
 
 function ProjectList() {
 
   return (
-    <div id="Projects">
+    <motion.div layout transition={{type: 'spring'}} id="Projects">
       <h2 id="project-card-title"> Projects</h2>
-      <div id="projects-list">
-        <div className="project-item">
+      <motion.div id="projects-list">
+        <motion.div className="project-item">
           <h4>CloudAEye</h4>
-        </div>
-        <div className="project-item">
+        </motion.div>
+        <motion.div className="project-item">
           <h4><Link to='FAM'>FAM</Link></h4>
-        </div>
-        <div className="project-item">
+        </motion.div>
+        <motion.div className="project-item">
           <h4>CloudAEye</h4>
-        </div>
-      </div>
-    </div>
+        </motion.div>
+      </motion.div>
+    </motion.div>
     
   )
 }
